Makefile: automate changelog & release
Changelog is automatically generated in a new file (prechangelog.rst),
to be edited and reviewed before final submission.

The version in ``.vimrc`` is also updated automatically according to the
new version being released.

Once this is done, changes are ready for commit and tag. The message for
these git objects should be the part of the changelog corresponding to
this version.

:smiley: